Improve naming for fluent builder API and extract components into their own files
