+++ api/src/main/kotlin/io/spine/protodata/renderer/SourceFile.kt
@@ -26,18 +26,13 @@
 
 package io.spine.protodata.renderer
 
-import com.google.common.base.Preconditions.checkArgument
 import com.google.common.base.Splitter
-import io.spine.logging.Logging
-import io.spine.protodata.InsertedPoint
 import io.spine.protodata.InsertedPoints
 import io.spine.protodata.filePath
 import io.spine.protodata.splitLines
 import io.spine.server.query.select
 import io.spine.text.Text
-import io.spine.text.TextFactory.checkNoSeparator
 import io.spine.text.TextFactory.text
-import io.spine.util.interlaced
 import java.lang.System.lineSeparator
 import java.nio.charset.Charset
 import java.nio.file.Path
@@ -122,8 +117,8 @@ private constructor(
      *
      * @see atInline
      */
-    public fun at(insertionPoint: InsertionPoint): SourceAtPoint =
-        SourceAtPoint(this, insertionPoint)
+    public fun at(insertionPoint: InsertionPoint): SourceAtLine =
+        SourceAtLine(this, insertionPoint)
 
     /**
      * Creates a new fluent builder for adding code at the given inline [insertionPoint].
@@ -145,9 +140,9 @@ private constructor(
             .orElseThrow()
         val point = points.pointList.filter { it.label == insertionPoint.label }.firstOrNull()
         return if (point != null) {
-            Located(this@SourceFile, point)
+            SpecificPoint(this@SourceFile, point)
         } else {
-            Unlocated
+            NoOp
         }
     }
 
@@ -274,162 +269,4 @@ private constructor(
     }
 
     override fun toString(): String = relativePath.toString()
-}
-
-/**
- * A fluent builder for inserting code into pre-prepared insertion points.
- *
- * @see SourceFile.at
- */
-public class SourceAtPoint
-internal constructor(
-    private val file: SourceFile,
-    private val point: InsertionPoint
-) {
-
-    private var indentLevel: Int = 0
-
-    /**
-     * Specifies extra indentation to be added to inserted code lines
-     *
-     * Each unit adds four spaces.
-     */
-    public fun withExtraIndentation(level: Int): SourceAtPoint {
-        checkArgument(level >= 0, "Indentation level cannot be negative.")
-        indentLevel = level
-        return this
-    }
-
-    /**
-     * Adds the given code lines at the associated insertion point.
-     *
-     * @param lines
-     *      code lines
-     */
-    public fun add(vararg lines: String): Unit =
-        add(lines.toList())
-
-    /**
-     * Adds the given code lines at the associated insertion point.
-     *
-     * @param lines
-     *      code lines
-     */
-    public fun add(lines: Iterable<String>) {
-        val sourceLines = file.lines()
-        val updatedLines = ArrayList(sourceLines)
-        val pointMarker = point.codeLine
-        val newCode = lines.linesToCode(indentLevel)
-        sourceLines.mapIndexed { index, line -> index to line }
-                   .filter { (_, line) -> line.contains(pointMarker) }
-                   .map { it.first + 1 }
-                   .forEach { index -> updatedLines.add(index, newCode) }
-        file.updateLines(updatedLines)
-    }
-}
-
-/**
- * A fluent builder for inserting code into pre-prepared inline insertion points.
- *
- * @see SourceFile.atInline
- */
-public sealed interface SourceAtPointInline {
-    /**
-     * Adds the specified code fragment into the insertion point.
-     *
-     * When called multiple times for the same insertion point, the code that is added last
-     * will appear first in the file, since new code fragments are always added right after
-     * the insertion point regardless of if it's been used before.
-     */
-    public fun add(codeFragment: String)
-}
-
-/**
- * A [SourceAtPointInline] at a point that could not be located.
- *
- * No code is generated by this implementation.
- */
-internal object Unlocated : SourceAtPointInline, Logging {
-    override fun add(codeFragment: String) = Unit
-}
-
-/**
- * A [SourceAtPointInline] at a point at a specific point or points in the file.
- *
- * The points are defined by the insertion point specified via the [point].
- */
-internal class Located(
-    private val file: SourceFile,
-    private val point: InsertedPoint
-) : SourceAtPointInline {
-
-    override fun add(codeFragment: String) {
-        checkNoSeparator(codeFragment)
-        val sourceLines = file.lines()
-        val updatedLines = ArrayList(sourceLines)
-        sourceLines.asSequence()
-            .mapIndexed { index, line -> CodeLine(index, line) }
-            .map { line -> line.insertInline(point, codeFragment) }
-            .forEach { (index, line) -> updatedLines[index] = line }
-        file.updateLines(updatedLines)
-    }
-}
-
-private fun String.splitByIndexes(indexes: List<Int>): List<String> = buildList {
-    val idxs = buildList(indexes.size + 2) {
-        add(0)
-        addAll(indexes)
-        add(length)
-    }
-    idxs.forEachIndexed { listIndex, stringIndex ->
-        if (listIndex < idxs.size - 1) {
-            val nextIndex = idxs[listIndex + 1]
-            add(substring(stringIndex, nextIndex))
-        }
-    }
-}
-
-private fun String.findInsertionIndexes(point: InsertedPoint): List<Int> {
-    val comment = point.representationInCode
-    var index = 0
-    return buildList {
-        while (true) {
-            val rawIndex = indexOf(comment, startIndex = index)
-            if (rawIndex >= 0) {
-                index = rawIndex + comment.length
-                add(index)
-            } else {
-                break
-            }
-        }
-    }
-}
-
-private fun Iterable<String>.linesToCode(indentLevel: Int): String =
-    joinToString(lineSeparator()) {
-        INDENT.repeat(indentLevel) + it
-    }
-
-private const val INDENT: String = "    "
-
-/**
- * A numbered line of code in a file.
- */
-private data class CodeLine(val lineIndex: Int, val content: String) {
-
-    /**
-     * Inserts the `newCode` at the given `insertionPoint` into this line and obtains the resulting
-     * line of code.
-     *
-     * The index of the new line is always the same as the index of the old line.
-     */
-    fun insertInline(point: InsertedPoint, newCode: String): CodeLine {
-        val indexes = content.findInsertionIndexes(point).toList()
-        if (indexes.isEmpty()) {
-            return this
-        }
-        val parts = content.splitByIndexes(indexes)
-        val newLine = parts.interlaced(newCode).joinToString(separator = "")
-        return CodeLine(lineIndex, newLine)
-    }
 }
